+++ test/chunking-form/samples/namespace-object-import/_config.js
@@ -1,0 +1,6 @@
+module.exports = {
+	description: 'namespace object import',
+	options: {
+		input: ['main1.js', 'main2.js']
+	}
+};