+++ test/chunking-form/samples/namespace-object-import/_expected/cjs/chunk1.js
@@ -1,0 +1,13 @@
+'use strict';
+
+var a = 'a';
+var b = 'a';
+
+var main2 = Object.freeze({
+	a: a,
+	b: b
+});
+
+exports.a = a;
+exports.b = b;
+exports.main2 = main2;