+++ test/chunking-form/samples/namespace-object-import/_expected/cjs/main1.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var __chunk1_js = require('./chunk1.js');
+
+console.log(__chunk1_js.a);
+
+console.log(__chunk1_js.main2);