+++ test/chunking-form/samples/namespace-object-import/_expected/es/main1.js
@@ -1,0 +1,5 @@
+import { main2, a, b } from './chunk1.js';
+
+console.log(a);
+
+console.log(main2);