+++ test/chunking-form/samples/namespace-object-import/_expected/system/main2.js
@@ -1,0 +1,16 @@
+System.register(['./chunk1.js'], function (exports, module) {
+	'use strict';
+	return {
+		setters: [function (module) {
+			var _setter = {};
+			_setter.a = module.a;
+			_setter.b = module.b;
+			exports(_setter);
+		}],
+		execute: function () {
+
+
+
+		}
+	};
+});