+++ test/chunking-form/samples/namespace-object-import/main1.js
@@ -1,0 +1,4 @@
+import * as main2 from './main2.js';
+import './dep.js';
+
+console.log(main2);