+++ test/form/samples/freeze/_expected/cjs.js
@@ -5,7 +5,6 @@ Object.defineProperty(exports, '__esModule', { value: true });
 const foo = 1;
 const bar = 2;
 
-
 var namespace = ({
 	foo: foo,
 	bar: bar