+++ test/form/samples/freeze/_expected/es.js
@@ -1,7 +1,6 @@
 const foo = 1;
 const bar = 2;
 
-
 var namespace = ({
 	foo: foo,
 	bar: bar