+++ test/form/samples/freeze/_expected/iife.js
@@ -4,7 +4,6 @@ var myBundle = (function (exports) {
 	const foo = 1;
 	const bar = 2;
 
-
 	var namespace = ({
 		foo: foo,
 		bar: bar