+++ test/form/samples/freeze/_expected/system.js
@@ -6,7 +6,6 @@ System.register([], function (exports, module) {
 			const foo = 1;
 			const bar = 2;
 
-
 			var namespace = ({
 				foo: foo,
 				bar: bar