+++ test/form/samples/legacy-reified-namespace/_expected/iife.js
@@ -4,7 +4,6 @@
 	const _typeof = 'typeof';
 	const foo = 1;
 
-
 	var namespace = (Object.freeze || Object)({
 		'typeof': _typeof,
 		foo: foo