+++ test/form/samples/legacy/_expected/cjs.js
@@ -3,7 +3,6 @@
 const foo = 1;
 const bar = 2;
 
-
 var namespace = (Object.freeze || Object)({
 	foo: foo,
 	bar: bar