+++ test/form/samples/legacy/_expected/es.js
@@ -1,7 +1,6 @@
 const foo = 1;
 const bar = 2;
 
-
 var namespace = (Object.freeze || Object)({
 	foo: foo,
 	bar: bar