+++ test/form/samples/legacy/_expected/system.js
@@ -6,7 +6,6 @@ System.register([], function (exports, module) {
 			const foo = 1;
 			const bar = 2;
 
-
 			var namespace = (Object.freeze || Object)({
 				foo: foo,
 				bar: bar