+++ test/form/samples/namespace-self-import/_expected/cjs.js
@@ -1,0 +1,13 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var self = Object.freeze({
+	get p () { return p$$1; }
+});
+
+console.log(Object.keys(self));
+
+var p$$1 = 5;
+
+exports.p = p$$1;