+++ test/form/samples/prefer-const/_expected/cjs.js
@@ -5,7 +5,6 @@ const other = require('other');
 const a = 1;
 const b = 2;
 
-
 const namespace = Object.freeze({
 	a: a,
 	b: b