+++ test/form/samples/prefer-const/_expected/system.js
@@ -10,7 +10,6 @@ System.register(['other'], function (exports, module) {
 			const a = 1;
 			const b = 2;
 
-
 			const namespace = Object.freeze({
 				a: a,
 				b: b